Convert csv to integration
