--- bf_opencore/mock/__init__.py
@@ -1,9 +1,0 @@
-# Copyright (C) 2017 Virta Laboratories, Inc.  All rights reserved.
-
-"""Mock asset connector.
-
-Creates mock assets
-"""
-
-from .__main__ import main, CONNECTOR_SPEC
-from .__main__ import DEFAULT_POPFILE
